Move the types codec out of the query codec class and file
